fixed change in copula package
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/spcopula@58 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/R/vineCopulas.R
@@ -18,7 +18,7 @@ validVineCopula = function(object) {
 }
 
 setClass("vineCopula",
-  representation = representation(copulas="list", dimension="numeric", type="character", pdf="numeric"),
+  representation = representation(copulas="list", dimension="integer", type="character", pdf="numeric"),
   validity = validVineCopula,
   contains = list("copula")
 )
